Un-hack schema rendering a bit
