+++ package.json
@@ -20,6 +20,7 @@
     "debug": "^2.1.3",
     "docker-exec-websocket-server": "^1.0.14",
     "express": "^4.13.4",
+    "got": "^8.2.0",
     "gulp": "^3.9.1",
     "gulp-awspublish": "git://github.com/djmitche/gulp-awspublish#noAcl",
     "gulp-clean": "^0.3.2",
@@ -28,6 +29,7 @@
     "gulp-file-include": "^1.1.0",
     "gulp-front-matter": "^1.3.0",
     "gulp-gunzip": "0.0.3",
+    "gulp-markdown": "^1.2.0",
     "gulp-rename": "^1.2.2",
     "gulp-untar": "0.0.4",
     "gulp-util": "^3.0.7",
@@ -37,7 +39,7 @@
     "jsdom": "^11.6.2",
     "json-schema-ref-parser": "^3.1.2",
     "lodash": "^4.11.1",
-    "marked": "^0.3.6",
+    "marked": "^0.3.17",
     "merge-stream": "^1.0.0",
     "mime": "^1.3.4",
     "morgan": "^1.7.0",
@@ -50,7 +52,6 @@
     "streamfilter": "^1.0.5",
     "superagent": "^1.1.0",
     "superagent-promise": "^0.2.0",
-    "sync-request": "^5.0.0",
     "taskcluster-client": "^2.1.0",
     "through2": "^2.0.1",
     "url-join": "^0.0.1",
